Fix race condition during start:dev
There's no way to run `aragon devchain` that makes it return to the
terminal with the chain running in the background. Since it never
returns, there's no good & quick way to run these two commands
back-to-back. Instead, this gets around the problem most of the time by
giving the devchain ten seconds to start up before continuing.